Assassin Manager Game
Includes the creation of two separate spaces of where characters can go: the kill ring and the graveyard. With the use of linked lists, player can enter a name of a character in the kill ring which will send them to the graveyard in that order, keeping track of the killer which is the character listed prior to the character selected in the linked list.